requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
watson-developer-cloud
simplejson
pandas
numpy


# ecdsa
# paramiko
# pycrypto
